% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alojamiento_cleaning.R
\name{alojamiento_cleaning}
\alias{alojamiento_cleaning}
\title{Alojamiento Cleaning}
\usage{
alojamiento_cleaning(base_datos)
}
\arguments{
\item{base_datos}{lista de las hojas que genero el full_reshape}
}
\value{
\itemize{
\item Un dataframe con la hoja principal, servicios y facilidades unificado por el _uuid
}

\itemize{
\item Ajustes de los nombres de las variables _uuid, _submission_time y _index
}

\itemize{
\item Ajustes los 1 y 0 por Si y No
}
}
\description{
Alojamiento Cleaning
}
\examples{
base_datos_reshape <- full_reshape_alojamiento(base_datos, etiquetas, choices)

base_cleaning <- cleaning_alojamiento(base_datos_reshape)
base_cleaning <- cleaning_alojamiento(base_datos_reshape = base_datos_reshape)
}
