% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alojamiento_merge.R
\name{alojamiento_merge}
\alias{alojamiento_merge}
\title{Alojamiento Merge}
\usage{
alojamiento_merge(base_datos)
}
\arguments{
\item{base_datos}{lista de las hojas que genero el full_reshape}
}
\value{
\itemize{
\item Un dataframe con la hoja principal, servicios y facilidades unificado por el _uuid
}

\itemize{
\item Ajustes de los nombres de las variables _uuid, _submission_time y _index
}

\itemize{
\item Ajustes los 1 y 0 por Si y No
}
}
\description{
Alojamiento Merge
}
\examples{
base_datos_reshape <- full_reshape_alojamiento(base_datos, etiquetas, choices)

base_merge <- alojamiento_merge(base_datos)
base_merge <- alojamiento_merge(base_datos = base_datos_reshape)
}
