% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_select_multiple.R
\name{change_select_multiple}
\alias{change_select_multiple}
\title{Change Select Multiple}
\usage{
change_select_multiple(encuesta, choices, survey, sep = "/")
}
\arguments{
\item{encuesta}{El parametro encuesta sera el dataframe de kobo}

\item{choices}{El parametro choices es la hoja Choices de la herramienta Kobo}

\item{survey}{El parametro survey es la hoja Survey de la herramienta Kobo}

\item{sep}{el separador de los select multiple "/", ".", "_"}
}
\value{
Un dataframe de la encuesta y con los valores de los select multiple convertidos
}
\description{
La funcion toma un dataframe de kobo y cambia los datos del select multiple 1 y 0
}
\examples{
survey <- read_excel("Input/KOBO_ABA_BOA VISTA - ANDRÉ_revGN_v6_GN.xlsx", sheet = "survey")
choices <- read_excel("Input/KOBO_ABA_BOA VISTA - ANDRÉ_revGN_v6_GN.xlsx", sheet = "choices")
df <-  read_excel("Input/REACH_BRA_ABAConsolidadoCleaning_20220906.xlsm", sheet = "Raw Data")

base <- change_select_multiple(df, choices, survey, "/")
base <- change_select_multiple(encuesta = df, choices = choices, survey = survey, sep = "/")

}
