% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercio_checks.R
\name{comercio_checks}
\alias{comercio_checks}
\title{Comercio Checks}
\usage{
comercio_checks(base_datos)
}
\arguments{
\item{base_datos}{La base de datos despues de haber realizado el cleaning}
}
\value{
InventariosLocos

PreciosLocos

OTROS

GISNombres

Feekback en excel por organizaciones
}
\description{
Sirve para identificar los datos atipicos de precios y valores atipicos de la base de datos
}
\examples{
base_cleaning <- cleaning_alojamiento(base_datos_reshape)

base_outliers <- checks_alojamiento(base_cleaning)
base_outliers <- checks_alojamiento(base_datos = base_cleaning)
}
