% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercio_final_cleaning.R
\name{comercio_final_cleaning}
\alias{comercio_final_cleaning}
\title{Comercio Final Cleaning}
\usage{
comercio_final_cleaning(df)
}
\arguments{
\item{df}{Dataframe final para realizar las correspondientes salidas en excel para HQ y GIS}
}
\value{
2 archivos de excel, uno para HQ y GIS

1 lista con los dos dataframe de HQ y GIS
}
\description{
Genera dos excel con la informacion correspondiente para HQ y GIS, tambien regresa una lista con estos dos dataframe
}
\examples{
lista_final <- comercio_final_cleaning(base_datos)
lista_final <- comercio_final_cleaning(df = base_datos)
}
