% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercio_full_reshape.R
\name{comercio_full_reshape}
\alias{comercio_full_reshape}
\title{Comercio Full Reshape}
\usage{
comercio_full_reshape(encuesta, etiquetas, choices)
}
\arguments{
\item{encuesta}{lista de las hojas que genero el kobo_to_r}

\item{etiquetas}{lista de las hojas que genero el import_rename}

\item{choices}{El parametro choices es la hoja Choices de la herramienta Kobo}
}
\value{
\itemize{
\item Una lista con los dataframe de la hoja principal, abastecimiento, food y non_food en reshape;
}

\itemize{
\item una lista con los nombres de las variables y su descripcion
}
}
\description{
Comercio Full Reshape
}
\examples{
choices <- read_excel("Input/REACH_JMMI_Survey.xlsx", 'choices')

base_datos <- kobo_to_r(directorio2, choices)

etiquetas <- import_raname(base_datos, choices, survey)

base_datos_reshape <- full_reshape(base_datos, etiquetas, choices)
base_datos_reshape <- full_reshape(base_datos = base_datos, etiquetas = etiquetas, choices = choices)
}
