% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercio_merge.R
\name{comercio_merge}
\alias{comercio_merge}
\title{Comercio Merge}
\usage{
comercio_merge(base_datos)
}
\arguments{
\item{base_datos}{lista de las hojas que genero el full_reshape}
}
\value{
\itemize{
\item Un dataframe con la hoja principal, abastecimiento, food y non_food unificado por el _uuid
}

\itemize{
\item Ajustes los 1 y 0 por Si y No
}
}
\description{
Comercio Merge
}
\examples{
base_datos_reshape <- full_reshape_alojamiento(base_datos, etiquetas, choices)

base_merge <- comercio_merge(base_datos)
base_merge <- comercio_merge(base_datos = base_datos_reshape)
}
