% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercio_replacements.R
\name{comercio_replacements}
\alias{comercio_replacements}
\title{Comercio Replacements}
\usage{
comercio_replacements(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
Dataframe con sin los numeros 888, 999 y sin las encuenesta donde no dieron el consentimiento
}
\description{
Variables con 999 o 888 las cambia a no sabe o se rehusa a responder
}
\examples{
output <- comercio_replacements(base_datos)
output <- comercio_replacements(df = base_datos)
}
