% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_separados.R
\name{excel_separados}
\alias{excel_separados}
\title{Excel Separados}
\usage{
excel_separados(df, variable = "organizacion")
}
\arguments{
\item{variable}{}
}
\value{

}
\description{
Genera salidas individuales segun la variable ingresada, ej. si la variable utilizada es
organizacion genera salida segun las organizaciones que se tengan
}
