% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_other_responses.R
\name{find_other_responses}
\alias{find_other_responses}
\title{Find Other Responses}
\usage{
find_other_responses(data, uuid = "uuid")
}
\arguments{
\item{data}{a dataframe}

\item{uuid}{como aparece el uuid en la base de datos}
}
\value{
Un dataframe con una fila por problema potencial. Tiene columnas para el índice de la fila correspondiente en los datos originales; el valor sospechoso; el nombre de la variable en el conjunto de datos original en el que se produjo el valor sospechoso; una descripción del tipo de problema.
}
\description{
Encontrar todas las respuestas en todas las columnas que podrían ser "especificar otras" respuestas a una pregunta de opción múltiple
}
\examples{
otros <- find_other_responses(base_datos)
}
