% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{Find Outliers}
\usage{
find_outliers(data, sd_max_p = 3, uuid = "uuid")
}
\arguments{
\item{data}{Un dataframe}

\item{sd_max_p}{Desviacion estandar, por default es 3, pero se puede ajustar}

\item{uuid}{Nombre de la variable uuid como se encuentra en la base de datos}
}
\value{
Un dataframe con una fila por problema potencial. Tiene columnas para el índice de fila correspondiente en los datos originales; el valor sospechoso; el nombre de la variable en el conjunto de datos original en el que se produjo el valor sospechoso; una descripción del tipo de problema.
}
\description{
Busca valores que estén fuera de más de tres desviaciones estándar de la media.
Si se encuentran menos valores atípicos cuando los datos se transforman en logaritmos antes de la comprobación, sólo se devuelven los valores atípicos en los datos transformados en logaritmos.
}
\examples{
outliers <- find_outliers(base_cleaning)
}
