% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rename.R
\name{import_rename}
\alias{import_rename}
\title{Import Rename}
\usage{
import_rename(encuesta, choices, survey)
}
\arguments{
\item{encuesta}{lista de las hojas que genero el kobo_to_r}

\item{choices}{El parametro choices es la hoja Choices de la herramienta Kobo}

\item{survey}{El parametro choices es la hoja Survey de la herramienta Kobo}
}
\value{
\itemize{
\item Una lista con los nombres de las variables y su descripcion
}
}
\description{
La funcion toma la salida generada con el kobo_to_r, de la herramienta kobo toma choices y survey para generar una lista con los nombre y etiquetas de la base de datos
}
\examples{
choices <- read_excel("Input/REACH_JMMI_Survey.xlsx", 'choices')
survey <- read_excel("Input/REACH_JMMI_Survey.xlsx", 'survey')

base_datos <- kobo_to_r(directorio2, choices)

etiquetas <- import_raname(base_datos, choices, survey)
etiquetas <- import_raname(base_datos = base_datos, choices = choices, survey = survey)
}
