% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.df_to_r.R
\name{kobo.df_to_r}
\alias{kobo.df_to_r}
\title{Kobo dataframe to R}
\usage{
kobo.df_to_r(encuesta, choices, survey, label_name = "label")
}
\arguments{
\item{encuesta}{El parametro encuesta sera el dataframe de kobo}

\item{choices}{El parametro survey es la hoja Survey de la herramienta Kobo}

\item{label_name}{El parametro label_name es el nombre de la etiqueta a la que se quiere transformar los datos}
}
\value{
\itemize{
\item Un dataframe de la encuesta y la convierte de numeros a los labels
}
}
\description{
La funcion toma un dataframe de kobo y lo cambia de name a label
}
\examples{
survey <- read_excel("Input/KOBO_ABA_BOA VISTA - ANDRÉ_revGN_v6_GN.xlsx", sheet = "survey")
choices <- read_excel("Input/KOBO_ABA_BOA VISTA - ANDRÉ_revGN_v6_GN.xlsx", sheet = "choices")
df <-  read_excel("Input/REACH_BRA_ABAConsolidadoCleaning_20220906.xlsm", sheet = "Raw Data")

base <- kobo.df_to_r(df, choices, survey, "label")
base <- kobo.df_to_r(encuesta = df, choices = choices, survey = survey, label_name = "label")
}
