% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_to_DAP.R
\name{kobo_to_DAP}
\alias{kobo_to_DAP}
\title{KOBO to DAP}
\usage{
kobo_to_DAP(sv, ch, label = "label")
}
\arguments{
\item{sv}{Datos de la hoja "survey" de la herramienta kobo}

\item{ch}{Datos de la hoja "choices" de la herramienta kobo}

\item{label}{Como aparece la variable label}
}
\value{
un dataframen con los datos para terminar de ajustar el DAP
}
\description{
Se alimenta de la información de la herramienta KOBO para generar un archivo DAP.
}
\examples{
datos <- Kobo_to_DAP(sv, ch, label = "label")
}
