% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_to_r.R
\name{kobo_to_r}
\alias{kobo_to_r}
\title{Kobo to R}
\usage{
kobo_to_r(encuesta, choices, survey)
}
\arguments{
\item{encuesta}{El parametro encuesta sera el directorio con el archivo descargado de kobo}

\item{choices}{El parametro survey es la hoja Survey de la herramienta Kobo}
}
\value{
\itemize{
\item Una lista con los dataframe de todas las hojas de la encuesta y la convierte de numeros a los labels
}
}
\description{
La funcion toma la encuesta descargada de kobo y la pasa a un dataframe
}
\examples{
name <- "REACH_JMMI_Survey.xlsx"
directorio <- paste("Input/", name, sep="")
choices <- read_excel(directorio, 'choices')

name2 <- "COL2001_JMMI_FECHA.xlsx"
encuesta <- paste("Input/", name2, sep="")

base <- kobo_to_r(encuesta, choices)
base <- kobo_to_r(encuesta = directorio2, choices = choices)
}
