% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosabe_noresponde.R
\name{nosabe_noresponde}
\alias{nosabe_noresponde}
\title{No sabe / No responde}
\usage{
nosabe_noresponde(df, uuid = "uuid")
}
\arguments{
\item{df}{Dataframe a verificar}

\item{uuid}{como aparece el uuui en el DF}
}
\value{
Un dataframe con la informacion variable, uuid y valor del no sabe no responde
}
\description{
Manera de encontrar los no sabe y no responde cuando las preguntas son de tipo númerico y se han utilizado los codígos 888 y 999
}
\examples{
nainteger <- nosabe_noresponde(df, uuid = "uuid")


}
