% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_labels.R
\name{obtain_labels}
\alias{obtain_labels}
\title{Obtain Labels}
\usage{
obtain_labels(sv, cc)
}
\arguments{
\item{sv}{el archivo survey de la herramienta de Kobo}

\item{cc}{el archivo choices de la herramienta de Kobo}
}
\value{
un dataframe con los names y sus debidas etiquetas
}
\description{
Obtain Labels
}
\examples{
etiquetas <- obtain_labels(survey, choices)
}
