% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_excel.R
\name{r_excel}
\alias{r_excel}
\title{Read Excel}
\usage{
r_excel(filename, tibble = FALSE)
}
\arguments{
\item{filename}{direccion del archivo que se quiere cargar}

\item{tibble}{}
}
\value{
una lista del archivo excel con todas las hojas
}
\description{
Genera una lista con las hojas del excel cargado
}
\examples{
separate_excel("C:/Users/SIG/Dropbox/REACH_COL/2_Research_Management/2022_41XXX_ECHO/MSNA/1_Setup/5_Training/4_Registro_encuestadores/Correo_registro/Kobo_MSNA2022_JM_V5.xlsx")
}
