% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alojamiento_replacements.R
\name{alojamiento_replacements}
\alias{alojamiento_replacements}
\title{Alojamiento Replacements}
\usage{
alojamiento_replacements(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
Dataframe con sin los numeros 888 y 999
}
\description{
Variables con 999 o 888 las cambia a no sabe o se rehusa a responder
}
\examples{
output <- replacements_alojamiento(base_datos)
output <- replacements_alojamiento(df = base_datos)
}
