% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incorporate_logs.R
\name{incorporate_logs}
\alias{incorporate_logs}
\title{Incorporate Logs}
\usage{
incorporate_logs(
  raw_df,
  cleaning_log,
  df_group_seprator = "/",
  uuid_col = "_uuid"
)
}
\arguments{
\item{raw_df}{dataframe}

\item{cleaning_log}{dataframe de los cambios a realizar}

\item{uuid_col}{Como se encuentra escrito el uuid en la base de datos}
}
\value{
Lista con el archivo cleaned_df

Lista con el archivo master_cleaning_log

Lista con el archivo logs_not_in_rawDF

Lista con el archivo cleaning_log.appli

Lista con el archivo duplicate_logs
}
\description{
Sirve para realizar los ajustes a la base de datos segun el uuid, variable, valor anterior y valor nuevo. Tambien se debe tener en cuenta en el archivo de cleaning_log que los ajustes se realizaran a los que tengan 'si' en la variable changed
}
\examples{
incorprated_logs <- incorporate_logs(my_df, cleaning_log, uuid_col = "uuid")
}
